+++ Staccato-iOS/Staccato-iOS/Util/Model/Memory/MemoriesResponse.swift
@@ -12,7 +12,7 @@ struct MemoriesResponse: Codable {
 }
 
 struct MemoryCandidateResponse: Codable {
-    let memoryId: Int64      // 메모리의 고유 ID
+    let memoryId: Int      // 메모리의 고유 ID
     let memoryTitle: String  // 메모리의 제목
     let startAt: String      // 메모리가 시작된 날짜와 시간
     let endAt: String        // 메모리가 끝난 날짜와 시간
